+++ src/TestSuite/TestCase.php
@@ -25,6 +25,8 @@ use Tools\TestSuite\TestTrait;
 
 /**
  * TestCase class
+ * @property string $alias The alias name for which a test is being performed
+ * @property class-string $originClassName The class name for which a test is being performed
  */
 abstract class TestCase extends CakeTestCase
 {
@@ -35,6 +37,36 @@ abstract class TestCase extends CakeTestCase
      * @var array
      */
     protected array $_cache = [];
+
+    /**
+     * Get magic method.
+     *
+     * It provides access to the cached properties of the test.
+     * @param string $name Property name
+     * @return mixed
+     * @since 2.23.0
+     * @throws \ReflectionException
+     * @throws \PHPUnit\Framework\AssertionFailedError
+     */
+    public function __get(string $name)
+    {
+        switch ($name) {
+            case 'alias':
+                if (empty($this->_cache['alias'])) {
+                    $this->_cache['alias'] = $this->getAlias($this->originClassName);
+                }
+
+                return $this->_cache['alias'];
+            case 'originClassName':
+                if (empty($this->_cache['originClassName'])) {
+                    $this->_cache['originClassName'] = $this->getOriginClassNameOrFail($this);
+                }
+
+                return $this->_cache['originClassName'];
+            default:
+                $this->fail('Property `' . $name . '` does not exist');
+        }
+    }
 
     /**
      * Called before every test method