+++ tests/test_app/TestApp/Test/TestCase/View/AppViewTest.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace App\Test\TestCase\View;
+
+use MeTools\TestSuite\TestCase;
+
+class AppViewTest extends TestCase
+{
+}